Fix loading of notification plugins
Inconsistent naming of plugins can cause an import error. This
commit standardises the naming so that all imports are parsed
correctly.

It also adds support for using short names for the previously
'built in plugins', which now need to be explictly enabled
when using Oslo config.

If this is confusing, then yes it is, the whole thing is a bit
of a mess.

Story: 2005545
Task: 30689

Change-Id: Ife27fed83d28d47cc99ee07c4a8c0c4dac32c2da